Move docker functions to a docker_lib module.
