+++ tests/test_parse.py
@@ -8,8 +8,8 @@ class TestParse(unittest.TestCase):
     def setUp(self):
         pass
 
-    def test_raw_jats(self):
+    def test_get_raw_jats(self):
         file_name = data_path('Dutzler 39122 edit.docx')
         expected = read_fixture('raw_jats_dutzler_39122.xml')
-        jats_content = parse.raw_jats(file_name)
+        jats_content = parse.get_raw_jats(file_name)
         self.assertEqual(jats_content, expected)
